fix windows build

+++ src/apps/snacman/snacman/simulations/snacgame/scene/GameScene.cpp
@@ -211,7 +211,7 @@ std::optional<Transition> GameScene::update(const snac::Time & aTime, RawInput &
 
     mSystems.get(update)->get<system::LevelCreator>().update();
 
-    mSystems.get(update)->get<system::PlayerInvulFrame>().update(aTime.mDeltaSeconds);
+    mSystems.get(update)->get<system::PlayerInvulFrame>().update((float)aTime.mDeltaSeconds);
     mSystems.get(update)->get<system::AllowMovement>().update();
     mSystems.get(update)->get<system::ConsolidateGridMovement>().update((float)aTime.mDeltaSeconds);
     mSystems.get(update)->get<system::IntegratePlayerMovement>().update((float)aTime.mDeltaSeconds);
@@ -223,11 +223,11 @@ std::optional<Transition> GameScene::update(const snac::Time & aTime, RawInput &
     mSystems.get(update)->get<system::SceneGraphResolver>().update();
 
     mSystems.get(update)->get<system::PortalManagement>().postGraphUpdate();
-    mSystems.get(update)->get<system::PowerUpUsage>().update(aTime.mDeltaSeconds);
+    mSystems.get(update)->get<system::PowerUpUsage>().update((float)aTime.mDeltaSeconds);
     mSystems.get(update)->get<system::EatPill>().update();
 
     mSystems.get(update)->get<system::RoundMonitor>().update();
-    mSystems.get(update)->get<system::PlayerSpawner>().update(aTime.mDeltaSeconds);
+    mSystems.get(update)->get<system::PlayerSpawner>().update((float)aTime.mDeltaSeconds);
 
     mSystems.get(update)->get<system::Debug_BoundingBoxes>().update();
